% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illdth-cif-data-analysis.r
\name{cifcmp.full}
\alias{cifcmp.full}
\title{Cumulative Incidence Function Estimation}
\usage{
cifcmp.full(
  age_recr,
  age_diag,
  age_death,
  age_end_fu,
  status_end,
  tgrd,
  tgrd.cb,
  auxflg,
  covpr,
  nresam,
  cmb.flg,
  adpwt,
  cmb.wgt,
  t.opt,
  pt.ci,
  bwts
)
}
\arguments{
\item{age_recr}{age at recruitment vector}

\item{age_diag}{age at disease diagnosis vector}

\item{age_death}{age at death vector}

\item{age_end_fu}{age at end of follow-up (death or censoring) vector (note that occurrence of disease does not end follow-up)}

\item{status_end}{status at end of follow-up vector (0=alive without disease, 1=died without disease, 2=alive with disease, 3=died with disease)}

\item{tgrd}{grid of ages at which the CIF will be computed (vector)}

\item{tgrd.cb}{grid of ages over which the simultaneous confidence band will be computed (vector)}

\item{auxflg}{logical flag for including (TRUE) or not including (FALSE) auxiliary terms}

\item{covpr}{desired confidence interval coverage probability}

\item{nresam}{number of bootstrap replications for confidence band}

\item{cmb.flg}{logical flag for including (TRUE) or not including (FALSE) combination estimator}

\item{adpwt}{logical flag for using (TRUE) or not using (FALSE) adaptive weights}

\item{cmb.wgt}{vector with length of length(tgrd) with weights for combination estimator (ignored if adpwt=TRUE)}

\item{t.opt}{option for transformation: 1=no transformation, 2=-log(1-u), 3=0.5*pi-arcsin(sqrt(1-u))}

\item{pt.ci}{option for pointwise CI's: 1=normal-theory, 2=bootstrap}

\item{bwts}{option for bootstrap weights: 1=normal, 2=exponential, 3=Poisson}
}
\value{
a list containing CIF estimates, SD estimates, pointwise confidence limits, simultaneous confidence band limits
}
\description{
Function to implement the cumulative incidence function (CIF) estimator
of Gorfine, Zucker, and Shoham for the illness-death model
}
